+++ HISTORY.md
@@ -1,5 +1,17 @@
 # Racing+ Version History and News
 
+### *v0.34.5* - May 3rd, 2019
+
+* The streak text will now always appear when you use a pill.
+* In a seeded race or a season 6 speedrun, it is no longer possible to become revived from Guppy's Collar (to prevent the vanilla permadeath bug).
+* Removed Broken Ankh from the trinket pool for seeded races and season 6.
+* If The Soul enters a trapdoor or beam of light, the game will now automatically switch to The Forgotten (to prevent bugs with teleporting the body).
+* If The Soul enters a room with Mom, Gurdy, or It Lives!, the game will now automatically switch to The Forgotten (to prevent bugs with teleporting the body).
+* Fixed the bug where The Soul would not switch back to the Forgotten if they died in a seeded race / season 6 speedrun. (Thanks NorBro86)
+* Fixed the bug where some rooms were incorrectly flipped. (Thanks thereisnofuture)
+* Fixed the bug where duplicate items could be started in season 6. (Thanks NorBro86)
+* Fixed the bug where the death timer was overlapping with the "S6" icon in season 6.
+
 ### *v0.34.4* - May 3rd, 2019
 
 * The logo for season 6 is improved. Thanks to BMZ_Loop for working on this.